requirements.txt
click>=8.0.0
pyyaml>=6.0
aiohttp>=3.8.0
aiofiles>=0.8.0
python-dotenv>=0.19.0
openai>=1.52.0